rsc_table_parser: Move header to make internal functions internal only
The files in include/openamp are deployed on install and can be used
by applications using the OpenAMP API. The functions in rsc_table_parser
are all internal use only. Move this header out of that location.
